Move TransactionTypeHelper
