+++ src/Extension/StateTransitions.php
@@ -9,6 +9,7 @@ use Wirecard\Order\State\State;
 /**
  * Interface StateTransitions
  * @package Wirecard\Order\State\Extension
+ * @todo name
  */
 interface StateTransitions
 {
