+++ src/TransactionType/Failure.php
@@ -3,6 +3,7 @@
 
 namespace Wirecard\Order\State\TransactionType;
 
+use Wirecard\Order\State\Implementation\TransactionTypeHelper;
 use Wirecard\Order\State\TransactionType;
 
 class Failure implements TransactionType
